refactor: edit form wrapper component, replace whitespaceTests.ts to helpers folder
