+++ src/shared/ui/form/formText/FormText.module.scss
@@ -1,0 +1,5 @@
+.signUpText {
+  font-size: var(--font-size-middle);
+  font-weight: var(--font-weight-bold);
+  margin-top: 20px;
+}
