+++ src/shared/ui/form/index.tsx
@@ -1,0 +1,4 @@
+import { FormWrapper } from './formWrapper/FormWrapper';
+import { FormText } from './formText/FormText';
+
+export { FormText, FormWrapper };
